frontend/streamlit/pyproject.toml
[project]
name = "streamlit-frontend"
version = "0.1.0"
description = "Streamlit frontend for GenAI application"
authors = [
    {name = "Nuttee", email = "nuttea@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.31.1",
    "httpx>=0.26.0",
    "requests>=2.31.0",
    "Pillow>=10.2.0",
    "pandas>=2.1.4",
    "python-dotenv>=1.0.0",
    "ddtrace>=3.18.0",  # Datadog LLMObs SDK for dataset management (requires 3.18+ for LLMObs features)
    "google-cloud-bigquery[pandas]>=3.25.0",  # BigQuery client for querying SS5/18 file metadata
    "google-genai>=1.0.0",  # Gemini API client for PDF extraction
    "pydantic>=2.0.0",  # Data validation models for election form data
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.1.1",
    "ruff>=0.1.14",
]

[tool.hatch.build.targets.wheel]
only-include = ["app.py", "pages/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = [
    "E402",   # Module level import not at top (Streamlit uses st.set_page_config first)
    "E501",   # Line too long (handled by Black for code, strings are okay)
    "N806",   # Variable naming (environment variables are uppercase by convention)
    "UP006",  # Type annotation style (not critical)
    "UP035",  # Deprecated typing imports (not critical for Python 3.11)
]
